+++ Nuake/src/Scripting/Modules/InputModule.h
@@ -47,8 +47,8 @@ namespace Nuake {
 
 			static void IsMouseButtonDown(WrenVM* vm)
 			{
-				int key = wrenGetSlotDouble(vm, 1);
-				bool result = Input::IsMouseButtonDown(key);
+				double key = wrenGetSlotDouble(vm, 1);
+				bool result = Input::IsMouseButtonDown((int)key);
 				wrenSetSlotBool(vm, 0, result);
 			}
 
@@ -61,28 +61,28 @@ namespace Nuake {
 
 			static void IsMouseButtonReleased(WrenVM* vm)
 			{
-				int key = wrenGetSlotDouble(vm, 1);
+				int key = (int)wrenGetSlotDouble(vm, 1);
 				bool result = Input::IsMouseButtonReleased(key);
 				wrenSetSlotBool(vm, 0, result);
 			}
 
 			static void IsKeyDown(WrenVM* vm)
 			{
-				int key = wrenGetSlotDouble(vm, 1);
+				int key = (int)wrenGetSlotDouble(vm, 1);
 				bool result = Input::IsKeyDown(key);
 				wrenSetSlotBool(vm, 0, result);
 			}
 
 			static void IsKeyPressed(WrenVM* vm)
 			{
-				int key = wrenGetSlotDouble(vm, 1);
+				int key = (int)wrenGetSlotDouble(vm, 1);
 				bool result = Input::IsKeyPressed(key);
 				wrenSetSlotBool(vm, 0, result);
 			}
 
 			static void IsKeyReleased(WrenVM* vm)
 			{
-				int key = wrenGetSlotDouble(vm, 1);
+				int key = (int)wrenGetSlotDouble(vm, 1);
 				bool result = Input::IsKeyReleased(key);
 				wrenSetSlotBool(vm, 0, result);
 			}
